+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Latest
 ------
+* Changed APILoader.get_location() so it always returns a valid Location.
+  Previously, if a location did not exist it created a placeholder with only the
+  identifier set. It now uses the ebird-api's new get_location() method to always
+  returns a valid Location.
 
 0.7.0 (2025-01-31)
 ------------------